Replace shard versions structure in Partition with a Map
The old list was not a good design - you could only add a new shard that
had a higher rowId; required expensive validation logic;  and the addShard API
required a version range which didn't fit ingestion too well. The new design
uses a Map which is more flexible and idempotent, and much safer.