Fix edge case row iteration skipping bugs
If the skiplist of the next chunk has row 0 in it, the skip logic would not get invoked because it comes before the next chunk logic,
and after that next() is called.  Swap the order so that next chunk is taken cared of first and then skips are checked, so first row
skips are properly taken cared of.
Also fix the case where an entire chunk has been skipped