Changing ShardMapper implementation to use the 'spread' concept inste… (#59)
* Changing ShardMapper implementation to use the 'spread' concept instead of N-Bits concept.

Spread is easier to understand. S is the 'spread' assigned for a given appName.
The data for every metric in the app is spread across 2^S shards. Example: if S=2,
data is spread across 4 shards. If S=0, data is located in 1 shard. Bigger apps
are assigned bigger S and smaller apps are assigned small S.