fix(coord): Removing Stale Coordinators after recovery of Cluster Singleton
Earlier attempt to recover singleton state had an issue because MemberRemoved
for the downed singleton node was never consistently delivered due to a race
condition with singleton failover event.

This fix removes the reliance on that event and removes stale coordinators
by comparing coordinators in the mappers to the cluster membership after
recovery. The coords that are present in the mappers, but not in the new
membership are unassigned.