bug(coord, core): Fix ser/deser of QueryError and exceptions
It so happens that IllegalStateException inside QueryError when deserialized initializes a null suppressedExceptions unmodifiable collection field. When kryo tries to serialize this exception again (happens in multiple propagations of results), we see a null pointer exception.

Fix is to explicitly provide a different serializer for unmodifiable collections list. This comes from a new open source library with extra serializers.