bug(core): Some partitions were indexed with later startTime on recovery
Since kafka recovery skips samples flushed to Cassandra, new partitions that
are not added to any index time bucket before restart, but whose samples are
already began flushing to C* happen to be indexed with later startTime during
recovery.

Fix is to create the partition if necessary for skipped rows without ingesting the
sample. This way partition gets created and index entry gets added for the right
time stamp.