bug(core,coord): More error logging from thread-pools
We have seen during testing when errors occur in the ingestion
pipeline without trace in logs. Shards have gone down without
an ability to identify root cause. Uncaught exception handlers
write to stdout and stderr which may not be captured in log
systems.

This commit configures all our thread pools with a custom
uncaught exception handler which logs the exception to log files.