feat(coord): Automatically Reassign error shard (#172)
Currently, when a shard errors out, it stays in error state. We do
not automatically restart the shards.

With this commit, when shard results in an IngestionError event,
ShardManager attempts to reassign it another node. Reassignment
is attempted only if last reassignment is more than 2 hours earlier.
This is so that we do not keep bouncing the shard for errors which
are more related to data, than the node itself.