bug(kafka): Replace monix-kafka usage in downsampling Kafka producer
No need to use queues and monix-kafka for downsampling since
Kafka producer takes care of the batching.

To use monix-kafka producer, it is necessary for the message source
to be transformed into an Observable. The queues were a workaround 
or that, but it is not necessary anymore since Kafka driver takes care
of the batching.