bug(query): mapAsync of child nodes causes incorrect order of child results (#261)
One of the recent changes to execute child plans in parallel
with mapAsync(numProcs) has caused regression in Binary Join.
Since LHS and RHS results can arrive different order, results are
sometimes incorrect. Fix is to keep track of original order of
ExecPlan nodes and use the order when composing the results.