bug(core): ODP from Cassandra only if Lucene index indicates that there is data (#291)
Prior to this commit, if earliestTime in memory for a partition is
more than queryStart time, we decide to page from cassandra.
This results in unnecessary calls to Cassandra because there
may not be any data.

With this change, we page from cassandra only if Lucene index
indicates that partition's startTime is earlier than the earliest
time in memory, and the query is for earlier than the earliest
time in memory.

The code organization for ODP IMO has been less than ideal
because there is a lot of things happening and leak in abstraction
with non-functional side effects. I do understand everything
has been done to reduce object allocation and double iteration
of objects. I have tried to make some improvements to existing code,
but we need to do a round of refactoring and cleanup to make
code more understandable.