bug(core): endTime update for re-ingesting time series should happen immediately (#357)
For intermittently ingesting time series, the endTime gets updated
as Long.MaxValue only during index time bucket flush event that
happens once in a flush interval. This is too late and queries prior
to this event skip this time series in result.

Fix is to check if series is not actively ingesting and update the
index immediately after re-ingestion. Since bitmap check is expensive,
we are adding an 'ingesting' flag per partition.