feat(various): Assert thread name during various operations (#441)
Currently there is a lot of confusion and un-clarity and risk of bugs
because the thread pool various operations get executed on does
not stand out in code so much so that it is almost impossible for a
new developer to confidently change code without risking introducing
new issues with concurrency.

This is an interim solution for introducing cleanliness, clarity and
maintainability by clearly conveying to reader/developer which
thread pool code is executed in.

Using this assertion, I was also able to fix a couple of bugs where 
* index recovery was being executed on akka actor thread instead of ingestion thread.
* removePartition was not happening on ingestion thread

Also bringing clarity in threadpool by making implicitly passed
scheduler arguments explicit. Also renamed ec and sched thread 
pools to ioPool in many places. Former names are too ambiguous 
and unclear and can lead to bugs.