feat(query, core): QueryResult need not always be serializable (#451)
Currently QueryResult uses SerializableRangeVectors only. With the new InProcessDispatcher which is coming soon, QueryResult can be a result of an ExecPlan that is executed within the current JVM. So we should not be forced into converting results into a SerializableRangeVector.

Hence QueryResult will use RangeVector and not force use of SerializableRangeVector.