bug(core): Release IndexSearchers after each use (#453)
After each acquire of searcher, we need to call release so that searcher
releases the segments and they are deleted after merge. Prior to this
change, release calls were not being made.

Currently, we are seeing a constant growth of disk space because old
segments are never deleted. Expecting that this change will fix it.

Also renamed commit operation to refresh. Those two are technically
different in Lucene. We are actually not doing commit at the bottom
of the call stack.