feat(core): Support for Binary Vectors on heap (#530)
Prior to this commit we were able to support Binary Vectors as native pointers only. This commit allows reading of Binary Vectors from on-heap too.

A MemoryAccessor abstraction is added that allows for reading from writing into memory. Different implementations provide accessors that read/write relative to different types of memory regions. The NativePointerAccessor is a singleton and allows native pointer access without incurring additional on-heap allocations. On-Heap accessors do incur an additional on-heap object to store base and offset-logic, but tradeoff here is the typed access in code which maintains sanity of having clean abstractions.

Binary Vectors now use accessor and address in the readers. Appenders are still off-heap. Support for on-heap appenders will be planned in another PR.

Also introduced ChunkSetInfoReader a trait used by the query engine to access Chunk Info records. We now have different implementations for on/off heap ChunkSetInfos. ChunkQueryInfo is removed and replaced with ChunkSetInfoReader implementation.

This PR enables many important features:
* Chunks can be read from Cassandra without incurring additional copy into off-heap memory
* No need for complexity adding native memory managers when dealing with encoded chunks. This is important for simplicity of applications such as downsamplers, job based non-real-time query engine, export to or read from object store like S3.
* Paves a path for type safe memory access