feat(sparkJobs): Cardinality Buster to handle ingestion of bad data (#709)
Bad data will be ingested - it is a certainty. If the bad data is extreme high cardinality
(like ip addresses and uuids assigned to tags), it will affect performance of the downsample
index. This spark job helps with deleting bad entries from the PartitionKeys table based on a
filter passed as job config. It deletes from downsample tables by default. But can can
configured to delete from raw tables as well. Use with care since it deletes data from the tables!