bug(core): Populate ODPed chunks in reverse order (#799)
OPDed chunks are populated oldest first. This can potentially 
cause another query to assume that chunks are already there 
in memory because it checks the oldest chunk time in the chunk-map.

This PR makes population of chunks in newest first order. This way concurrent 
queries will take the ODP path. Not efficient since ODP will be done 
twice but, acceptable for now. At least it will yield right query results.