feat(query): Introducing RangeVectorCursor (#804)
Today RangeVectors expose sample rows as Iterator[RowReader] which cannot be closed.
Because of this, there isn't a cascading way to close the ChunkInfoIterator, thereby 
causing room for developer error where locks are not released. 

This PR changes RangeVector to use RangeVectorCursor which is designed to 
cascade the close call to all wrapped iterators, finally culminating with close of 
the ChunkInfoIterator(s), thereby releasing the locks reliably. This PR will force 
developers during compile time to think about closing their underlying cursors.