perf(query): Remove logging of plans on error to reduce logging I/O (#972)
We use printTree that logs plan on errors. Now that we have original promQL query,
we dont need to log full exec plan - we can log the query. ExecPlan logging is huge,
esp with spread=8. This reduces logging pressure. We found that in some nodes
that are backed on the scheduler, a lot of time is spent logging timeout errors