bug(core): Fix Partition Eviction Issues (#1095)
Currently, we use lucene index to identify partIds eligible for eviction.
This is not proving to be reliable. Getting top-k partitions with oldest
endTime from lucene is complex and is complex possibly has bugs.
Even if fixed, it yields to complex eviction logic. There are cases when
ODP may cause the partition to never be evicted at all because of
existing watermark logic.

Hence:
* Explicitly maintain a queues of partIds eligible for eviction. When
  partition stops ingesting or is ODPed, the partId is added to this list.
  Eviction happens very quickly from this queue. ODPed partitions
  get evicted first to be consistent with block eviction.
* Another issue is we still have ODP cannibalism where ODP queries
  can silently evict other TSPs required by the same ODP query. This
  is prevented by disallowing inline eviction for ODP queries.
* We create partition headroom just like in Block memory, and evict
  time series partitions regularly to make room for ODP partitions.  
* Holding eviction write lock prevents races between eviction and queries
  and pins partitions to shard until queries complete.