bug(prometheus): Enforce maximum regex length (#1305)
Upon very long regular expression filters, Lucene returns a StackOverflowError
on every shard, and this has the potential to degrade performance of the
cluster severely. The shard lock is not released and subsequent queries
fail. The StackOverflowError has also been seen on ElasticSearch.

The stack overflow error we saw was for a regex filter of length 17k
characters, and it was an unintentional regex which was not per intent.
It used comma instead of pipe character to delimit list of possibilities.

Since it is very hard to dive into the regex to explicitly allow and disallow
only certain regexes causing performance issues, we have to resort to
enforcing a regular expression filter length limit - currently set to 1KB,
which is same as ElasticSearch default. We will iteratively visit improving
this strategy.