fix(exec): override LocalPartitionDistConcatExec schema reducer (#1408)
When child ResultSchemas are reduced for a LocalPartitionDistConcatExec, the default reducer is used. This throws a SchemaMismatch exception if the schemas aren't identical (except for some nuance when empty schemas are involved). This reduction strategy is incompatible with BinaryJoin pushdowns, where child joins might reduce schemas differently. This commit relaxes the schema-equality constraint in the same way as other DistConcat types.