feat(query): QueryStats now includes CPU time used by query (#1512)
Query Statistics now includes CPU time spent on query. This can be used to do analytics and rate limit on queries.

CPU Time is calculated in these places:
* Step1 of exec plan execution to construct the monix query pipeline. This is where index lookup happens for leaf plans.
* Consumption of the pipelined Observables at plan result serialization time (step2). This will cover for all queries including aggregations and pipelined transformations.
* Binary Join and Set operation compose since they load all data to memory and this operation is not pipelined
* Top/BottomK Result serialization

The above list is not comprehensive, hence the number reported is not complete. We do not want delays from execution of other queries to impact CPU time measurement of the current query. The main challenge in achieving completeness is that stream processing of data makes it nasty to accurately calculate CPU time since it would be sum of time for all pipelined lambdas. This will require us to spew instrumentation all over the place. For now, we place it at places where we see significant usage.