feat(query): Query Planning optimization for 'or vector(0)' (#1530)
foo or vector(0) like queries are very common. Current plan materialization results in SetOperatorExec which is expensive since it transmits data over the network and uses more memory.

A more optimal implementation would be to apply an InstantFunctionMapper to convert the NaNs to 0s. This plan is memory efficient and does not write results to the wire unnecessarily. This is what this PR delivers.

PR description includes results of 1-node performance test.