fix(core): Bug in calculating size of SerializedRangeVector (#1643)
There was a bug in calculating size of SRV.
Earlier, for efficiency purposes, we were calculating the size of the containers associated with the SRV.
But actually, the container can home multiple SRVs. So the calculated size for several SRVs at a time can end up wrong with addition of cumulative counts.

The fix for now is to calculate the size by going through the records. It introduces a small inefficiency here, but submitting this PR for now since other ways to calculate this were more invasive and risk regression. We can have an optimization of this if really needed later. I have also reduced the number of calls to this method from two to one.

The unit tests didn't catch this since earlier since they played with one SRV only.
I have now added a unit test that adds multiple SRVs. It failed with earlier code.