feat(core): Add open/close support for Tantivy Index (#1817)
Commit 1 of adding Tantivy index support.  This PR is broken down into a small subset
of the overall Tantivy logic to allow for easier partial review.  The index code is not
usable end to end until all parts are committed.

This adds the basic Rust project skeleton and supports opening, applying schema, and closing the index.
Many methods are unimplemented and will be added in follow up PRs.  End to end testing is not available
in this PR as the index test suite requires ingestion and query support for verification.