feat(core): Add ingestion support for Tantivy
This implements enough of the interface to add documents to the Tantivy
index.  This is part 2 of the series and is not usable end to end until
additional PRs complete.

Document additional and removal is covered in this PR.  Documents are
created by calling a single native method that has many of the built-in fields,
such as start and end time, passed as method parameters.  Dyanmic fields
are passed to the Rust code via a simple binary encoding to reduce
JNI overhead.