feat(core): Add Scala side of Tantivy query support
This contains the Scala side of query support.  This completes the basic end to
end flow, so it also includes regression and performance tests.

There will be some small cleanup PRs after this for things like additional metrics,
but with this PR the Tantivy index can be successfully activated via config.