feat(core): Optimize Rust code by default
Previously the code only optimized if a flag was passed during sbt invocation
or if we're building multi-arch.  This has created confusing results
where perf drifts because debug mode was used.

After doing testing and development it was found that release is a suitable
default - it does incremental build, includes symbols for debugging,
and while not as fast as debug builds is still fast enough for an inner loop.

You can still build with debug by setting rust.optimize=false for any
cases where it is needed during local development.