fix(build): Support glibc qualifiers for Rust targets (#1858)
If you're targetting an older Linux distro the default glibc version
being linked against may be too high to produce runnable images.

cargo zigbuild supports specifying a specific glibc version to use
for a link target by appending ".<version>" to the target triple.  For
example, "x86_64-unknown-linux-gnu.2.17" will target v2.17.

For the most part this just works but we need to strip this suffix when
looking for output binaries.  This fix adds that logic.