fix(core): Fix tantivy column cache not releasing memory from deleted segments (#1864)
Columns in the column cache hold a reference to the mmaped file data that backs
the segment.  These segments can be deleted during segment merging, but if a column
for that segment is in the cache it prevents the mmap from closing and releasing RAM.

To fix this we subscribe for notifications on segment list changes and clear the
column cache when these occur so stale segments can be reclaimed.