fix(core): Error downsampled query if lookback is short when querying rate/increase (#1882)
Rate requires 2 samples in lookback window. Instead of returning empty result for rate in downsampled dataset (such queries with small window and smaller publish interval are encountered frequently in downsampled dataset), we error the query out when lookback is insufficient. This fix is imperfect, but serves as a workaround until a better solution is delivered. Status quo affects customers by producing wrong query results and leads to thinking that there is data loss, or even worse, implicit trust that the query results are right.