Revert "feat(coord): raw-data query with range expression also route to downSample cluster (#1947)" (#1961)
Stitching a RawSeries plan of raw and downsampled data is not supported due to schema differences:

Raw data schema has a column of (value),
Downsampled data schema has a column of (avg).
When stitching this data, it tries to reduce schema and throws SchemaMismatch Error.
Reverting this commit to remove new logic that may stitch raw/downsampled RawSeries data.